Version 4.2.2

+++ wp-admin/includes/upgrade.php
@@ -524,11 +524,14 @@ function upgrade_all() {
 	if ( $wp_current_db_version < 29630 )
 		upgrade_400();
 
+	// Don't harsh my mellow. upgrade_422() must be called before
+	// upgrade_420() to catch bad comments prior to any auto-expansion of
+	// MySQL column widths.
 	if ( $wp_current_db_version < 31534 )
 		upgrade_422();
 
-	if ( $wp_current_db_version < 31536 )
-		upgrade_423();
+	if ( $wp_current_db_version < 31351 )
+		upgrade_420();
 
 	maybe_disable_link_manager();
 
@@ -1423,6 +1426,19 @@ function upgrade_400() {
  * @since 4.2.0
  */
 function upgrade_420() {
+	global $wp_current_db_version, $wpdb;
+
+	if ( $wp_current_db_version < 31351 && $wpdb->charset === 'utf8mb4' ) {
+		if ( is_multisite() ) {
+			$tables = $wpdb->tables( 'blog' );
+		} else {
+			$tables = $wpdb->tables( 'all' );
+		}
+
+		foreach ( $tables as $table ) {
+			maybe_convert_table_to_utf8mb4( $table );
+		}
+	}
 }
 
 /**
@@ -1477,31 +1493,6 @@ function upgrade_422() {
 
 		foreach ( $comments as $comment ) {
 			wp_delete_comment( $comment->comment_ID, true );
-		}
-	}
-}
-
-/**
- * Execute changes made in WordPress 4.2.0.
- *
- * @since 4.2.3
- */
-function upgrade_423() {
-	global $wp_current_db_version, $wpdb;
-
-	if ( $wp_current_db_version < 31536 && $wpdb->charset === 'utf8mb4' ) {
-		if ( is_multisite() ) {
-			$tables = $wpdb->tables( 'blog' );
-		} else {
-			$tables = $wpdb->tables( 'all' );
-			if ( defined( 'DO_NOT_UPGRADE_GLOBAL_TABLES' ) ) {
-				$global_tables = $wpdb->tables( 'global' );
-				$tables = array_diff_assoc( $tables, $global_tables );
-			}
-		}
-
-		foreach ( $tables as $table ) {
-			maybe_convert_table_to_utf8mb4( $table );
 		}
 	}
 }
@@ -1605,7 +1596,7 @@ function upgrade_network() {
 
 	// 4.2
 	if ( $wp_current_db_version < 31351 && $wpdb->charset === 'utf8mb4' ) {
-		if ( ! defined( 'DO_NOT_UPGRADE_GLOBAL_TABLES' ) ) {
+		if ( ! ( defined( 'DO_NOT_UPGRADE_GLOBAL_TABLES' ) && DO_NOT_UPGRADE_GLOBAL_TABLES ) ) {
 			$wpdb->query( "ALTER TABLE $wpdb->usermeta DROP INDEX meta_key, ADD INDEX meta_key(meta_key(191))" );
 			$wpdb->query( "ALTER TABLE $wpdb->site DROP INDEX domain, ADD INDEX domain(domain(140),path(51))" );
 			$wpdb->query( "ALTER TABLE $wpdb->sitemeta DROP INDEX meta_key, ADD INDEX meta_key(meta_key(191))" );
@@ -1621,7 +1612,7 @@ function upgrade_network() {
 
 	// 4.2.2
 	if ( $wp_current_db_version < 31535 && 'utf8mb4' === $wpdb->charset ) {
-		if ( ! defined( 'DO_NOT_UPGRADE_GLOBAL_TABLES' ) ) {
+		if ( ! ( defined( 'DO_NOT_UPGRADE_GLOBAL_TABLES' ) && DO_NOT_UPGRADE_GLOBAL_TABLES ) ) {
 			$upgrade = false;
 			$indexes = $wpdb->get_results( "SHOW INDEXES FROM $wpdb->signups" );
 			foreach( $indexes as $index ) {
@@ -1633,17 +1624,6 @@ function upgrade_network() {
 
 			if ( $upgrade ) {
 				$wpdb->query( "ALTER TABLE $wpdb->signups DROP INDEX domain_path, ADD INDEX domain_path(domain(140),path(51))" );
-			}
-		}
-	}
-
-	// 4.2.3
-	if ( $wp_current_db_version < 31536 && $wpdb->charset === 'utf8mb4' ) {
-		if ( ! defined( 'DO_NOT_UPGRADE_GLOBAL_TABLES' ) ) {
-			$tables = $wpdb->tables( 'global' );
-
-			foreach ( $tables as $table ) {
-				maybe_convert_table_to_utf8mb4( $table );
 			}
 		}
 	}
@@ -1777,17 +1757,6 @@ function maybe_convert_table_to_utf8mb4( $table ) {
 				return false;
 			}
 		}
-	}
-
-	$table_details = $wpdb->get_row( "SHOW TABLE STATUS LIKE '$table'" );
-	if ( ! $table_details ) {
-		return false;
-	}
-
-	list( $table_charset ) = explode( '_', $table_details->Collation );
-	$table_charset = strtolower( $table_charset );
-	if ( 'utf8mb4' === $table_charset ) {
-		return true;
 	}
 
 	return $wpdb->query( "ALTER TABLE $table CONVERT TO CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci" );
@@ -1949,9 +1918,6 @@ function dbDelta( $queries = '', $execute = true ) {
 	 */
 	$iqueries = apply_filters( 'dbdelta_insert_queries', $iqueries );
 
-	$text_fields = array( 'tinytext', 'text', 'mediumtext', 'longtext' );
-	$blob_fields = array( 'tinyblob', 'blob', 'mediumblob', 'longblob' );
-
 	$global_tables = $wpdb->tables( 'global' );
 	foreach ( $cqueries as $table => $qry ) {
 		// Upgrade global tables only for the main site. Don't upgrade at all if DO_NOT_UPGRADE_GLOBAL_TABLES is defined.
@@ -2021,24 +1987,9 @@ function dbDelta( $queries = '', $execute = true ) {
 
 				// Is actual field type different from the field type in query?
 				if ($tablefield->Type != $fieldtype) {
-					$do_change = true;
-					if ( in_array( strtolower( $fieldtype ), $text_fields ) && in_array( strtolower( $tablefield->Type ), $text_fields ) ) {
-						if ( array_search( strtolower( $fieldtype ), $text_fields ) < array_search( strtolower( $tablefield->Type ), $text_fields ) ) {
-							$do_change = false;
-						}
-					}
-
-					if ( in_array( strtolower( $fieldtype ), $blob_fields ) && in_array( strtolower( $tablefield->Type ), $blob_fields ) ) {
-						if ( array_search( strtolower( $fieldtype ), $blob_fields ) < array_search( strtolower( $tablefield->Type ), $blob_fields ) ) {
-							$do_change = false;
-						}
-					}
-
-					if ( $do_change ) {
 					// Add a query to change the column type
-						$cqueries[] = "ALTER TABLE {$table} CHANGE COLUMN {$tablefield->Field} " . $cfields[strtolower($tablefield->Field)];
-						$for_update[$table.'.'.$tablefield->Field] = "Changed type of {$table}.{$tablefield->Field} from {$tablefield->Type} to {$fieldtype}";
-					}
+					$cqueries[] = "ALTER TABLE {$table} CHANGE COLUMN {$tablefield->Field} " . $cfields[strtolower($tablefield->Field)];
+					$for_update[$table.'.'.$tablefield->Field] = "Changed type of {$table}.{$tablefield->Field} from {$tablefield->Type} to {$fieldtype}";
 				}
 
 				// Get the default value from the array
@@ -2495,9 +2446,14 @@ function pre_schema_upgrade() {
 		}
 	}
 
+	if ( $wp_current_db_version < 30133 ) {
+		// dbDelta() can recreate but can't drop the index.
+		$wpdb->query( "ALTER TABLE $wpdb->terms DROP INDEX slug" );
+	}
+
 	// Upgrade versions prior to 4.2.
 	if ( $wp_current_db_version < 31351 ) {
-		if ( ! is_multisite() && ! defined( 'DO_NOT_UPGRADE_GLOBAL_TABLES' ) ) {
+		if ( ! is_multisite() ) {
 			$wpdb->query( "ALTER TABLE $wpdb->usermeta DROP INDEX meta_key, ADD INDEX meta_key(meta_key(191))" );
 		}
 		$wpdb->query( "ALTER TABLE $wpdb->terms DROP INDEX slug, ADD INDEX slug(slug(191))" );
